BFBBFix Version 1.3
(1.2 were the two commits that were pushed on Aug 12th)

Electricity Generators in Museum now deal damage just as the ones in Mermalair do. I've been looking for instances of objects with the same models but different behaviors but may not have found all of them. If you know of any please let me know.

Removed an invisible barrier in Museum that normally disappears when approaching it from the second level. Removing it doesn't impact casual players' necessity to find a way to disable/bypass the laser walls, so it was determined to be unecessary movement blocking for the sake of preventing confusion. Such confusion has been deemed to be rather miniscule, as the worst case scenario is the player getting near the end of the map and realizing they have to disable the laser field to progress. While a pretty bad outcome, you'd likely have to be new to both BFBB and 3D platformers in general to encounter it, which is unlikely given the target demographic of this mod.

Removed the last two dialogues with the Mermalair Computer in favour of adding a warp back to the ballroom after unlocking the villain containment center. This results in the loss of a reminder text and some fluff text where Spongebob greets the Computer without response. While the loss of reminder text is normally not great, it may be misleading considering it rather vaguely reminds the player to do something in another map, a purpose the warp menu serves in a less confusing fashion.

Covered the tip of a rock below a previously added Shh Flying Tiki Combo in Sand Mountain Hub with a OOB trigger. This should prevent the player from getting stuck down there and forcing them to jump off into the OOB trigger below. Basically it's a small time saving/anti-frustration measure.

Expanded the Mermalair Computer's forced dialogue sphere to cover the entirety of the platform to prevent a bug that prevents players from ever being able to push the security shutdown button if they push the button next to the computer after the other three security buttons and before talking to the Mermalair Computer.

Fixed the pressed variant of the master shutdown button being improperly rotated.